+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add support for building connector images from the Maven coordinates
 * Allow Kafka Connect Build artifacts to be downloaded from insecure servers (#5542)
 * Configurable authentication, authorization, and SSL for Cruise Control API
+* Update to Cruise Control version 2.5.73
 
 ### Changes, deprecations and removals